% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_meta_species_format.R
\name{check_meta_species_format}
\alias{check_meta_species_format}
\title{Check the 'format' of species metaphlan tables}
\usage{
check_meta_species_format(df)
}
\arguments{
\item{df}{dataframe to be checked.}
}
\value{
format, a character value equal to:
\enumerate{
\item 'rel_perc' stand for percentage relative abundances.
\item 'rel' stand for non percentage relative abundances.
\item 'count' stand for counts data.
}
}
\description{
Internal function that check whether the metaphlan derived table is in count,
relative abundance or percentage relative abundance. These checks are simply
made based on rowSums and colSums expected values.
}
