% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_meta_species_table.R
\name{check_meta_species_table}
\alias{check_meta_species_table}
\title{Check metaphlan derived species datasets}
\usage{
check_meta_species_table(df)
}
\arguments{
\item{df}{Species dataset derived from metaphlan profile}
}
\description{
The internal function checks whether the dataset is subset to species level through the
's__' pattern searched in all column names .
}
