% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_sampleid_column.R
\name{check_sampleid_column}
\alias{check_sampleid_column}
\title{Check sampleid column in metaphlan derived table}
\usage{
check_sampleid_column(df)
}
\arguments{
\item{df}{Dataframe to inspect.}
}
\description{
Internal function that check if a sampleid column is present based on popular
choices of names for this kind of data. If the column is indeed present it is
removed from the dataset and returned in the list. On the contrary a NULL value
is reported.
}
