% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divide_numeric_cols.R
\name{divide_numeric_cols}
\alias{divide_numeric_cols}
\title{Split dataframe in numeric and non numeric columns}
\usage{
divide_numeric_cols(df)
}
\arguments{
\item{df}{Dataframe to be splitted.}
}
\description{
Internal function used for splitting a dataframe in a list of 2 dataframes:
one made of numeric columns and the other made of other types columns.
}
