% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_zero_cols.R
\name{filter_zero_cols}
\alias{filter_zero_cols}
\title{Filter numeric columns containing only zero values}
\usage{
filter_zero_cols(df)
}
\arguments{
\item{df}{Dataframe to filter.}
}
\value{
The function returns the filtered dataframe.
}
\description{
The function filter the numeric columns of only zero values. The non-numeric
columns are not touched and collected at beginning of the dataset.
}
