% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_presence_perc_cat.R
\name{get_presence_perc_cat}
\alias{get_presence_perc_cat}
\title{Get the percentage presence of all numeric columns for levels of a factor.}
\usage{
get_presence_perc_cat(
  df,
  cat_var,
  min_abn_value = 0,
  min_abn_equal = FALSE,
  perc_on_all = FALSE,
  ...
)
}
\arguments{
\item{df}{dataframe with numeric columns and factor feature of interest.}

\item{cat_var}{Character value reporting the name of the column variable on which
levels the percentages of numeric columns are computed. Not-character columns are automatically
coerced to factors.}

\item{min_abn_value}{Numeric value. Sometimes only abundance values greater than a minimum value
have to be considered as presence. This value is intended as a percentage value for microbial abundances
(e.g 5 means 5\% in relative abundances).}

\item{min_abn_equal}{Logical (default FALSE); if a value must be equal or greater (TRUE) or
strictly greater (FALSE) than the min_abn_value to be considered as a presence.}

\item{perc_on_all}{Logical (default FALSE); if the percentages must be computed on the total
number of rows/observations of the dataframe (TRUE) or on the number of observations of each group (FALSE).}

\item{...}{Needed to pass the additional argument 'do_check' that allows to block or not "format values checking",
useful in case of microbial abundances data.}
}
\value{
A dataframe with the numeric features in rows and the grouping levels on columns.
The cells report the percentage values.
}
\description{
The function allows to compute the percentage presences of all numeric columns,
in conjunction with a minimum cutoff of presence/absence, all in respect to the
levels of a factor variable.
}
