% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_species_richness.R
\name{get_species_richness}
\alias{get_species_richness}
\title{Get species richness index.}
\usage{
get_species_richness(
  df,
  margin = 2,
  min_abn_value = 0,
  min_abn_equal = FALSE,
  ...
)
}
\arguments{
\item{df}{dataframe of abundances data.}

\item{margin}{numeric; it can be set only equal to 1, case in which the microbial features are
on the rows, and 2 (default value), case in which the microbial features are on the columns.}

\item{min_abn_value}{Numeric value. Sometimes only abundance values greater than a minimum value
have to be considered as presence. This value is intended as a percentage value for microbial abundances
(e.g 5 means 5\% for relative abundances).}

\item{min_abn_equal}{logical (default FALSE); if a value must be equal or greater (TRUE) or
strictly greater (FALSE) than the min_abn_value to be considered a presence.}

\item{...}{Needed to pass additional argument do_check in order to block or not "format" values checking
useful in case of microbial abundances data.}
}
\value{
The function return a dataframe with the richness score plus the non-numeric columns
in case of margin = 2, or the names of columns in case of margin = 1.
}
\description{
The function allow to compute the species richness index.
}
\examples{


mpa_complete <- phlantools::mpa_complete
richness_df <- get_species_richness(df = mpa_complete, margin = 2)

## The function will return a message highlighting that also taxa not identified
## to species level are present and therefore used for the computation of the index.
## The function doesn't automatically select the species level taxa.
}
