% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpa_complete.R
\docType{data}
\name{mpa_complete}
\alias{mpa_complete}
\title{Metaphlan table with all taxonomies}
\format{
A data frame containing taxonomic information from Metaphlan analysis.
It contains 20 rows and 3741 columns:
\describe{
\item{sampleID}{Column with sample IDS}
\item{K_bacteria...}{Multiple columns reporting relative abundances of bacteria}
}
}
\usage{
mpa_complete
}
\description{
A subset of a metaphlan table generated using the tool on shotgun generated reads.
}
\keyword{datasets}
