% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpa_species.R
\docType{data}
\name{mpa_species}
\alias{mpa_species}
\title{Metaphlan table with only species level microorganisms}
\format{
A data frame containing species-level information from Metaphlan software.
It contains 20 rows and 200 columns, of which one report the samples id.
}
\usage{
mpa_species
}
\description{
A subset of a metaphlan table generated on shotgun reads.
}
\keyword{datasets}
