% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phlantools-package.R
\docType{package}
\name{phlantools-package}
\alias{phlantools}
\alias{phlantools-package}
\title{phlantools: Tools for performing common procedures on metagenomic "phlan" software-derived datasets}
\description{
This package contains functions that simplify the execution of common operations done on metagenomic "phlan" data (e.g. metaphaln profiles).
}
\keyword{internal}
