% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_functions.R
\name{check_functions}
\alias{check_functions}
\alias{check_species}
\alias{check_df}
\alias{check_sampleid_column}
\alias{check_table_format}
\title{Checks for metaphlan derived datasets}
\usage{
check_species(df)

check_df(df)

check_sampleid_column(df)

check_table_format(df)
}
\arguments{
\item{df}{dataframe to inspect.}
}
\description{
Set of internal functions to check whether the dataset is a dataframe, is subset to species level,
has a sampleid column and the format (respectively).
}
