% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_presence_table.R
\name{get_presence_table}
\alias{get_presence_table}
\title{Internal function to obtain a presence/absence table from an abundance table}
\usage{
get_presence_table(df, min_abn_value = 0, min_abn_equal = FALSE)
}
\arguments{
\item{df}{Dataframe from which obtain the presence/absence table.}

\item{min_abn_value}{numeric value. Sometimes only abundance values greater than a minimum value
have to be considered as presence. This value MUST be specified as percentage value (e.g 5 --> that means 5\%).}

\item{min_abn_equal}{logical (default FALSE); if a value must be equal or greater (TRUE) or
strictly greater (FALSE) than the min_abn_value to be considered a presence.}
}
\value{
The function returns the presence/absence table.
}
\description{
The function checks the format of the abundances table based simply on expectation
about its values and takes action accordingly.
}
