% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpa_complete.R
\docType{data}
\name{mpa_complete}
\alias{mpa_complete}
\title{Metaphlan table with all taxonomies}
\format{
A data frame containing taxonomic information from Metaphlan analysis.
It contains 20 rows and 100 columns, of which one report the samples id.
}
\usage{
mpa_complete
}
\description{
A subset of a metaphlan table generated on shotgun reads.
}
\keyword{datasets}
